Move BukkitFlyingProjectile
